dom/traversal/TreeWalker-navigate-from-filter.html

<!DOCTYPE html>
<html>
<!--
Test adapted from http://mxr.mozilla.org/mozilla-central/source/content/base/test/test_bug559526.html

https://bugzilla.mozilla.org/show_bug.cgi?id=559526
-->
<head>
<title>TreeWalker: navigate-from-filter</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="traversal-support.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<div id=log></div>
</head>
<body>
<p>Test TreeWalker and NodeIterator navigating from within the NodeFilter</p>
<div id="nodes" style="display:none">
  <div id="child1"></div>
  <div id="child2"></div>

  <div id="child3"></div>
  <div id="child4"></div>
  <div id="child5"></div>
  <div id="child6"></div>
  <div id="child7"></div>
  <div id="child8"></div>
</div>
<script>
var it;
var recurse = false;
var testCount;
function filter(node) {
  if (node.id == "child3" && ! recurse) {
    recurse = true;
    var ex = null;
    try {
      var foo = it.nextNode();
    } catch(e) {
      ex = e;
    }
    ++testCount;
    assert_equals(ex.name, "InvalidStateError", "Should have thrown an exception!");
    assert_equals(ex.code, DOMException.INVALID_STATE_ERR, "Should have thrown an exception!");
    recurse = false;
  }
  return /*NodeFilter.*/FILTER_ACCEPT;
}

test(function()
{
  testCount = 0;
  it = document.createNodeIterator(
    document.getElementById("nodes"),
    /*NodeFilter.*/SHOW_ELEMENT,
    filter
  );
  while(it.nextNode());
  assert_equals(testCount, 1);
}, 'Testing NodeIterator.nextNode');

test(function()
{
  testCount = 0;
  it = document.createTreeWalker(
    document.getElementById("nodes"),
    /*NodeFilter.*/SHOW_ELEMENT,
    filter
  );
  while(it.nextNode());
  assert_equals(testCount, 1);
}, 'Testing TreeWalker.nextNode');

test(function()
{
  testCount = 0;
  it = document.createTreeWalker(
    document.getElementById("nodes"),
    /*NodeFilter.*/SHOW_ELEMENT,
    filter
  );
  while(it.nextSibling());
  assert_equals(testCount, 1);
}, 'Testing TreeWalker.nextSibling');

test(function()
{
  testCount = 0;
  it = document.createTreeWalker(
    document.getElementById("nodes"),
    /*NodeFilter.*/SHOW_ELEMENT,
    filter
  );
  while(it.previousSibling());
  assert_equals(testCount, 1);
}, 'Testing TreeWalker.previousSibling');

// Cleanup
document.body.removeChild(document.getElementById("nodes"));

</script>
</body>
</html>
